[frontiers] Create service to check is a point is still a frontier (after exploration).
