+++ frontiers_msgs/srv/CheckIsFrontier.srv
@@ -1,0 +1,3 @@
+geometry_msgs/Point candidate
+---
+bool is_frontier
